Update Lakers to version of 12th May 2025 (#114)
The Rust version used by the GitHub actions CI cannot be used to compile Lakers v0.8.0 anymore.

Rather than fighting with how to downgrade Rust versions in the CI, this PR updates the commit SHA of Lakers to that of May 12, 2025 which solves that particular problem. Perhaps once a new tagged Lakers version is available, we should switch to that one.